requirements.txt
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.3
selenium==4.18.1
webdriver-manager==4.0.1
pyautogui==0.9.54 